ADD makefile, FIX compile errors with -Wall -Werror -Wextra in the changed files, GITIGNORE added objects files (.o) & library files (.a)

+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: drobert- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/15 16:24:54 by drobert-          #+#    #+#             */
-/*   Updated: 2022/02/15 16:43:10 by drobert-         ###   ########.fr       */
+/*   Updated: 2022/02/16 12:39:07 by drobert-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	*ft_memchr(const void *s, int c, size_t n)
 	int	i;
 
 	i = -1;
-	while (++i < n)
+	while (++i < (int)n)
 		if (*((unsigned char *)s + i) == (unsigned char)c)
 			return ((char *)s + i);
 	return (0);
