+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: drobert- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/14 15:17:33 by drobert-          #+#    #+#             */
-/*   Updated: 2022/02/14 16:01:20 by drobert-         ###   ########.fr       */
+/*   Updated: 2022/02/16 12:39:23 by drobert-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void	*ft_memcpy(void *dst, const void *src, size_t n)
 	i = -1;
 	dst_p = (char *)dst;
 	src_p = (char *)src;
-	while (++i < n)
+	while (++i < (int)n)
 		*(dst_p + i) = *(src_p + i);
 	return (dst);
 }
