+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: drobert- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/14 16:04:55 by drobert-          #+#    #+#             */
-/*   Updated: 2022/02/15 15:17:15 by drobert-         ###   ########.fr       */
+/*   Updated: 2022/02/16 12:39:40 by drobert-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void	*ft_memset(void *b, int c, size_t len)
 
 	fill = (unsigned char) c;
 	i = -1;
-	while (++i < len)
+	while (++i < (int)len)
 		*(unsigned char *)(b + i) = fill;
 	return (b);
 }
