+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: drobert- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/14 13:14:15 by drobert-          #+#    #+#             */
-/*   Updated: 2022/02/14 13:43:16 by drobert-         ###   ########.fr       */
+/*   Updated: 2022/02/16 12:40:06 by drobert-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t dst_size)
 	src_len = ft_strlen(src);
 	if (dst_size == 0)
 		return (src_len);
-	while (*src && ++n < dst_size - 1)
+	while (*src && ++n < (int)dst_size - 1)
 		*dst++ = *src++;
 	dst = 0;
 	return (src_len);
