+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: drobert- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/15 11:49:22 by drobert-          #+#    #+#             */
-/*   Updated: 2022/02/15 13:14:34 by drobert-         ###   ########.fr       */
+/*   Updated: 2022/02/16 12:55:43 by drobert-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 	i = 0;
 	if (n == 0)
 		return (0);
-	while (++i < n && *s1 && *s2 && *s1 == *s2)
+	while (++i < (int)n && *s1 && *s2 && *s1 == *s2)
 	{
 		s1++;
 		s2++;
