+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: drobert- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/15 11:30:59 by drobert-          #+#    #+#             */
-/*   Updated: 2022/02/15 11:45:50 by drobert-         ###   ########.fr       */
+/*   Updated: 2022/02/16 12:56:08 by drobert-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,10 +18,10 @@ char	*ft_strnstr(const char *haystack, const char *needle, size_t n)
 	int	j;
 
 	i = -1;
-	while (++i < n && *(haystack + i))
+	while (++i < (int)n && *(haystack + i))
 	{
 		j = 0;
-		while (*(needle + j) && *(haystack + i + j) && i + j < n
+		while (*(needle + j) && *(haystack + i + j) && i + j < (int)n
 			&& *(needle + j) == *(haystack + i + j))
 			j++;
 		if (*(needle + j) == 0)
